requirements.txt
black==25.1.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
flake8==7.2.0
iniconfig==2.1.0
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
PyYAML==6.0.2
tomli==2.2.1
typing-extensions==4.13.2
